renamed tabsPanel to SubjectDashboardTabs
