--- src/dataEntryApp/views/subjectDashBoard/components/TabsPanel.js
@@ -1,93 +1,0 @@
-import React, { Fragment } from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Tabs from '@material-ui/core/Tabs';
-import Tab from '@material-ui/core/Tab';
-import PropTypes from 'prop-types';
-import Typography from '@material-ui/core/Typography';
-import Paper from '@material-ui/core/Paper';
-import TabViewFirst from './TabView';
-import Box from '@material-ui/core/Box';
-
-function TabPanel(props) {
-    const { children, value, index, ...other } = props;
-
-    return (
-        <Typography
-            component="div"
-            role="tabpanel"
-            hidden={value !== index}
-            id={`scrollable-auto-tabpanel-${index}`}
-            aria-labelledby={`scrollable-auto-tab-${index}`}
-            {...other}
-        >
-            {value === index && <Box p={3}>{children}</Box>}
-        </Typography>
-    );
-}
-
-TabPanel.propTypes = {
-    children: PropTypes.node,
-    index: PropTypes.any.isRequired,
-    value: PropTypes.any.isRequired,
-};
-
-function a11yProps(index) {
-    return {
-        id: `scrollable-auto-tab-${index}`,
-        'aria-controls': `scrollable-auto-tabpanel-${index}`,
-    };
-}
-
-const useStyles = makeStyles(theme => ({
-
-    tabsDisplay: {
-        margin: theme.spacing(1)
-    },
-    tabView: {
-        // marginTop: theme.spacing(2),
-        backgroundColor: 'white',
-        boxShadow: 'none'
-    }
-}));
-
-export default () => {
-    const classes = useStyles();
-
-    const [value, setValue] = React.useState(0);
-
-    const handleChange = (event, newValue) => {
-        setValue(newValue);
-    };
-
-    return (
-        <Fragment>
-            <AppBar className={classes.tabView} position="static" color="default">
-                <Tabs
-                    value={value}
-                    onChange={handleChange}
-                    indicatorColor="primary"
-                    textColor="primary"
-                    variant="scrollable"
-                    scrollButtons="auto"
-                    aria-label="scrollable auto tabs example"
-                >
-                    <Tab label="Item One" {...a11yProps(0)} />
-                    <Tab label="Item Two" {...a11yProps(1)} />
-                    <Tab label="Item Three" {...a11yProps(2)} />
-                </Tabs>
-            </AppBar>
-            <TabPanel value={value} index={0}>
-                <Paper className={classes.tabsDisplay}>
-                    <TabViewFirst />
-                </Paper>
-            </TabPanel>
-            <TabPanel value={value} index={1}>
-                Item Two
-      </TabPanel>
-            <TabPanel value={value} index={2}>
-                Item Three
-      </TabPanel>
-        </Fragment>
-    );
-}
